+++ src/com/intellij/idea/plugin/hybris/settings/HybrisProjectSettings.kt
@@ -23,35 +23,37 @@ import com.intellij.idea.plugin.hybris.flexibleSearch.settings.FlexibleSearchSet
 import com.intellij.idea.plugin.hybris.groovy.settings.GroovySettings
 import com.intellij.idea.plugin.hybris.impex.settings.ImpexSettings
 import com.intellij.idea.plugin.hybris.polyglotQuery.settings.PolyglotQuerySettings
+import com.intellij.openapi.components.BaseState
 import java.util.*
 
-data class HybrisProjectSettings(
-    var flexibleSearchSettings: FlexibleSearchSettings = FlexibleSearchSettings(),
-    var polyglotQuerySettings: PolyglotQuerySettings = PolyglotQuerySettings(),
-    var impexSettings: ImpexSettings = ImpexSettings(),
-    var groovySettings: GroovySettings = GroovySettings(),
+class HybrisProjectSettings : BaseState() {
+    var flexibleSearchSettings: FlexibleSearchSettings = FlexibleSearchSettings()
+    var polyglotQuerySettings: PolyglotQuerySettings = PolyglotQuerySettings()
+    var impexSettings: ImpexSettings = ImpexSettings()
+    var groovySettings: GroovySettings = GroovySettings()
 
-    var customDirectory: String? = null,
-    var hybrisDirectory: String? = null,
-    var configDirectory: String? = null,
-    var importedByVersion: String? = null,
-    var hybrisVersion: String? = null,
-    var javadocUrl: String? = null,
-    var sourceCodeFile: String? = null,
-    var externalExtensionsDirectory: String? = null,
-    var externalConfigDirectory: String? = null,
-    var externalDbDriversDirectory: String? = null,
-    var ideModulesFilesDirectory: String? = null,
-    var hybrisProject: Boolean = false,
-    var importOotbModulesInReadOnlyMode: Boolean = false,
-    var followSymlink: Boolean = false,
-    var scanThroughExternalModule: Boolean = true,
-    var excludeTestSources: Boolean = false,
-    var importCustomAntBuildFiles: Boolean = false,
-    var completeSetOfAvailableExtensionsInHybris: Set<String> = mutableSetOf(),
-    var unusedExtensions: Set<String> = mutableSetOf(),
-    var modulesOnBlackList: Set<String> = mutableSetOf(),
+    var customDirectory by string(null)
+    var hybrisDirectory by string(null)
+    var configDirectory by string(null)
+    var importedByVersion by string(null)
+    var hybrisVersion by string(null)
+    var javadocUrl by string(null)
+    var sourceCodeFile by string(null)
+    var externalExtensionsDirectory by string(null)
+    var externalConfigDirectory by string(null)
+    var externalDbDriversDirectory by string(null)
+    var ideModulesFilesDirectory by string(null)
+    var hybrisProject by property(false)
+    var importOotbModulesInReadOnlyMode by property(false)
+    var followSymlink by property(false)
+    var scanThroughExternalModule by property(true)
+    var excludeTestSources by property(false)
+    var importCustomAntBuildFiles by property(false)
+    var showFullModuleName by property(false)
+    var completeSetOfAvailableExtensionsInHybris: Set<String> = mutableSetOf()
+    var unusedExtensions: Set<String> = mutableSetOf()
+    var modulesOnBlackList: Set<String> = mutableSetOf()
     var availableExtensions: MutableMap<String, ExtensionDescriptor> = TreeMap<String, ExtensionDescriptor> { a, b ->
         a.compareTo(b, true)
-    },
-)
+    }
+}